+++ tools/ess-nextgen-migration-toolkit/src/modules/postprocessing/__init__.py
@@ -1,1 +1,5 @@
-"""postprocessing package."""
+"""Postprocessing stage package."""
+
+from modules.postprocessing.output_pipeline import build_output_pipeline
+
+__all__ = ["build_output_pipeline"]